+++ src/commands/mod.rs
@@ -44,7 +44,6 @@ define_commands! {
         // Expression Commands
         EvaluateCommand,
         AssignCommand,
-        IncrementCommand,
 
         // Control flow commands
         IfCommand,